moved Create Course tests to tests.py
